Implemented DomainEvents.. Transitioned from Traditional to EventSourcing pattern

+++ VetClinic.Management.Domain/ValueObjects/WeightRange.cs
@@ -2,10 +2,10 @@
 {
     public record WeightRange
     {
-        public decimal From { get; set; }
-        public decimal To { get; set; }
+        public double From { get; set; }
+        public double To { get; set; }
 
-        public WeightRange(decimal from, decimal to)
+        public WeightRange(double from, double to)
         {
             From = from;
             To = to;
